refactor: move transport files to proxy/transport/ subdirectory
- transport.ts -> transport/http.ts
- stream-proxy.ts -> transport/stream.ts
- transport-fn.ts -> transport/transport-fn.ts
- Fixed all internal and test import paths
