+++ src/proxy/transport/http.ts
@@ -1,10 +1,10 @@
 import { request as httpRequestFn } from "http";
 import { request as httpsRequestFn } from "https";
-import { UPSTREAM_SUCCESS, filterHeaders } from "./types.js";
-import { buildUpstreamUrl } from "./proxy-core.js";
-import type { RawHeaders, TransportResult } from "./types.js";
+import { UPSTREAM_SUCCESS, filterHeaders } from "../types.js";
+import { buildUpstreamUrl } from "../proxy-core.js";
+import type { RawHeaders, TransportResult } from "../types.js";
 // Re-export callStream from stream-proxy.ts for external consumers
-export { callStream } from "./stream-proxy.js";
+export { callStream } from "./stream.js";
 
 // ---------- Constants ----------
